feat: voting page now in components
